requirements.txt
requests>=2.25
beautifulsoup4>=4.9
psycopg2-binary>=2.9
pandas
